refactor(desktop): split github card into sub-files
