+++ apps/desktop/src/features/github/components/Card/panes/CommentsPane.tsx
@@ -1,0 +1,140 @@
+import { useMemo, useState } from 'react';
+import { ExternalLink } from 'lucide-react';
+import type { PrComment, PullRequestState } from '@goodboy/types';
+import { groupThreads, threadPriority } from '../../../comment-threads';
+import { COMMENT_DISPLAY_LIMIT } from '../lib';
+import { CommentThreadRow } from './CommentThreadRow';
+
+type Props = {
+  readonly comments: ReadonlyArray<PrComment>;
+  readonly pr: PullRequestState;
+  readonly onOpenUrl: (url: string) => void;
+  readonly onSpawnFromComment?: (c: PrComment) => void;
+};
+
+export const CommentsPane = ({ comments, pr, onOpenUrl, onSpawnFromComment }: Props) => {
+  const [expanded, setExpanded] = useState<ReadonlySet<string>>(new Set());
+  const [showAll, setShowAll] = useState(false);
+  const [showResolved, setShowResolved] = useState(false);
+
+  const allThreads = useMemo(() => groupThreads(comments), [comments]);
+
+  const reviewThreads = useMemo(
+    () => allThreads.filter((t) => t.head.source === 'review'),
+    [allThreads],
+  );
+  const generalCount = allThreads.length - reviewThreads.length;
+  const resolvedCount = useMemo(
+    () => reviewThreads.filter((t) => t.head.resolved === true).length,
+    [reviewThreads],
+  );
+
+  const threads = useMemo(() => {
+    const filtered = showResolved
+      ? reviewThreads
+      : reviewThreads.filter((t) => t.head.resolved !== true);
+    return [...filtered].sort((a, b) => {
+      const p = threadPriority(a) - threadPriority(b);
+      if (p !== 0) {
+        return p;
+      }
+      return b.head.createdAt.localeCompare(a.head.createdAt);
+    });
+  }, [reviewThreads, showResolved]);
+
+  const generalFooter =
+    generalCount > 0 ? (
+      <button
+        type="button"
+        onClick={() => onOpenUrl(pr.url)}
+        className="inline-flex items-center gap-0.5 text-[10px] text-muted-foreground/70 hover:text-foreground"
+        title="open general comments on GitHub"
+      >
+        {generalCount} general comment{generalCount === 1 ? '' : 's'}
+        <ExternalLink size={9} aria-hidden />
+      </button>
+    ) : null;
+
+  if (reviewThreads.length === 0) {
+    return (
+      <div className="flex flex-col gap-1 text-[11px] text-muted-foreground">
+        <span>No review comments yet</span>
+        {generalFooter}
+      </div>
+    );
+  }
+
+  if (threads.length === 0) {
+    return (
+      <div className="flex flex-col gap-1 text-[11px] text-muted-foreground">
+        <div className="flex items-center gap-1.5">
+          <span>All review comments resolved 🎉</span>
+          {resolvedCount > 0 && (
+            <button
+              type="button"
+              onClick={() => setShowResolved(true)}
+              className="text-[10px] underline-offset-2 hover:text-foreground hover:underline"
+            >
+              show {resolvedCount}
+            </button>
+          )}
+        </div>
+        {generalFooter}
+      </div>
+    );
+  }
+
+  const visible = showAll ? threads : threads.slice(0, COMMENT_DISPLAY_LIMIT);
+  const hidden = threads.length - visible.length;
+
+  const toggle = (id: string) => {
+    setExpanded((prev) => {
+      const next = new Set(prev);
+      if (next.has(id)) {
+        next.delete(id);
+      } else {
+        next.add(id);
+      }
+      return next;
+    });
+  };
+
+  return (
+    <ul className="flex flex-col gap-1.5">
+      {visible.map((t) => (
+        <li key={t.head.id}>
+          <CommentThreadRow
+            thread={t}
+            expanded={expanded.has(t.head.id)}
+            onToggle={() => toggle(t.head.id)}
+            onOpenUrl={onOpenUrl}
+            onSpawn={onSpawnFromComment ? () => onSpawnFromComment(t.head) : undefined}
+          />
+        </li>
+      ))}
+      {hidden > 0 && (
+        <li>
+          <button
+            type="button"
+            onClick={() => setShowAll(true)}
+            className="text-[10px] text-muted-foreground hover:text-foreground"
+          >
+            +{hidden} more
+          </button>
+        </li>
+      )}
+      {resolvedCount > 0 && (
+        <li>
+          <button
+            type="button"
+            onClick={() => setShowResolved((v) => !v)}
+            className="text-[10px] text-muted-foreground/70 hover:text-foreground"
+          >
+            {showResolved ? `hide ${resolvedCount} resolved` : `show ${resolvedCount} resolved`}
+          </button>
+        </li>
+      )}
+      {generalFooter ? <li>{generalFooter}</li> : null}
+    </ul>
+  );
+};
