+++ apps/desktop/src/features/github/components/Card/parts/DetailSkeleton.tsx
@@ -1,0 +1,9 @@
+export const DetailSkeleton = () => {
+  return (
+    <div className="flex flex-col gap-1.5" role="status" aria-label="loading pr data">
+      <div className="h-2.5 w-3/4 animate-pulse rounded bg-muted [animation-delay:0ms]" />
+      <div className="h-2.5 w-1/2 animate-pulse rounded bg-muted [animation-delay:120ms]" />
+      <div className="h-2.5 w-2/3 animate-pulse rounded bg-muted [animation-delay:240ms]" />
+    </div>
+  );
+};
