+++ apps/desktop/src/features/github/components/Card/parts/StaleCaption.tsx
@@ -1,0 +1,25 @@
+import { useNow } from '../../../../../shared/hooks/useNow';
+import { formatRelative } from '../lib';
+
+type Props = {
+  readonly fetchedAt: string | null;
+};
+
+export const StaleCaption = ({ fetchedAt }: Props) => {
+  const now = useNow(30_000, !!fetchedAt);
+  if (!fetchedAt) {
+    return null;
+  }
+  const ageMs = now - new Date(fetchedAt).getTime();
+  if (!Number.isFinite(ageMs) || ageMs < 60_000) {
+    return null;
+  }
+  return (
+    <span
+      className="text-[9px] text-muted-foreground/60"
+      title={`fetched at ${new Date(fetchedAt).toLocaleString()}`}
+    >
+      updated {formatRelative(ageMs)}
+    </span>
+  );
+};
